feat(sdk): let @package/agent use the @paradym/wallet-sdk
