+++ packages/agent/src/batch.ts
@@ -1,10 +1,12 @@
 import { Mdoc, MdocRecord, SdJwtVcRecord, W3cCredentialRecord } from '@credo-ts/core'
+import {
+  getBatchCredentialMetadata,
+  getCredentialCategoryMetadata,
+  getRefreshCredentialMetadata,
+} from '@package/sdk/src/metadata/credentials'
+import { decodeW3cCredential } from '@package/sdk/src/utils/encoding'
 import { RefreshPidUseCase } from '../../../apps/easypid/src/use-cases/RefreshPidUseCase'
 import type { EitherAgent } from './agent'
-import { getCredentialCategoryMetadata } from './credentialCategoryMetadata'
-import { decodeW3cCredential } from './format/credentialEncoding'
-import { getBatchCredentialMetadata } from './openid4vc/batchMetadata'
-import { getRefreshCredentialMetadata } from './openid4vc/refreshMetadata'
 import { updateCredential } from './storage'
 
 export async function refreshPid({
