+++ packages/agent/src/display/category.ts
@@ -1,6 +1,6 @@
 import { ClaimFormat } from '@credo-ts/core'
 import { getPidAttributesForDisplay } from '@easypid/utils/pidCustomMetadata'
-import type { CredentialCategoryMetadata } from '../credentialCategoryMetadata'
+import type { CredentialCategoryMetadata } from '@package/sdk/src/metadata/credentials'
 
 const categoryDisplay = {
   [ClaimFormat.SdJwtVc]: {
