+++ packages/agent/src/openid4vc/registerDcApi.ts
@@ -1,12 +1,12 @@
 import { type RegisterCredentialsOptions, registerCredentials } from '@animo-id/expo-digital-credentials-api'
 import { DateOnly, type Logger, type MdocNameSpaces } from '@credo-ts/core'
+import { getCredentialForDisplay } from '@package/sdk/src/display/credential'
 import { sanitizeString } from '@package/utils'
 import * as ExpoAsset from 'expo-asset'
 import { Image } from 'expo-image'
 import { ImageManipulator, SaveFormat } from 'expo-image-manipulator'
 import { Platform } from 'react-native'
 import type { EitherAgent } from '../agent'
-import { getCredentialForDisplay } from '../display'
 
 type CredentialItem = RegisterCredentialsOptions['credentials'][number]
 type CredentialDisplayClaim = NonNullable<CredentialItem['display']['claims']>[number]
