+++ packages/app/src/components/CredentialCard.tsx
@@ -1,5 +1,4 @@
-import type { DisplayImage } from '@package/agent'
-
+import type { DisplayImage } from '@package/sdk/src/display/credential'
 import {
   Card,
   Heading,
