+++ packages/app/src/features/notifications/components/PresentationNotificationScreen.tsx
@@ -1,5 +1,3 @@
-import type { FormattedSubmission } from '@package/agent'
-
 import {
   Button,
   FloatingSheet,
@@ -15,6 +13,7 @@ import { sanitizeString } from '@package/utils'
 import { useEffect, useState } from 'react'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
+import type { FormattedSubmission } from '@package/sdk/src/format/submission'
 import { useNavigation } from 'expo-router'
 import { CredentialRowCard, DualResponseButtons } from '../../../components'
