+++ packages/sdk/package.json
@@ -29,7 +29,9 @@
 		"@credo-ts/didcomm": "catalog:",
 		"@credo-ts/openid4vc": "catalog:",
 		"@credo-ts/react-native": "catalog:",
-		"@openid4vc/openid4vp": "catalog:"
+		"@openid4vc/openid4vp": "catalog:",
+		"@sd-jwt/decode": "catalog:",
+		"@sd-jwt/present": "catalog:"
 	},
 	"peerDependencies": {
 		"react": "catalog:",
