+++ pnpm-workspace.yaml
@@ -17,9 +17,7 @@ catalog:
   babel-plugin-module-resolver: ^5.0.2
 
   # Credo / Agent / SSI
-  credo-ts-didweb-anoncreds: 0.0.1-alpha.14
-  rxjs: ^7.8.1
-  query-string: ^8.1.0
+  "@astronautlabs/jsonpath": ^1.1.2
   "@animo-id/eudi-wallet-functionality": 0.0.0-alpha-20250604122408
   "@animo-id/expo-ausweis-sdk": 0.0.1-alpha.14
   "@animo-id/expo-secure-environment": 0.1.0
@@ -32,8 +30,6 @@ catalog:
   "@openid4vc/openid4vp": 0.3.0-alpha-20250602121005
   "@openid4vc/openid4vci": 0.3.0-alpha-20250602121005
   "@openid4vc/utils": 0.3.0-alpha-20250602121005
-  isomorphic-webcrypto: npm:@sphereon/isomorphic-webcrypto
-  fast-text-encoding: ^1.0.6
   "@types/fast-text-encoding": ^1.0.3
   "@credo-ts/anoncreds": npm:@animo-id/credo-ts-anoncreds@0.5.14-alpha-20250604111322
   "@credo-ts/askar": npm:@animo-id/credo-ts-askar@0.5.14-alpha-20250604111322
@@ -48,8 +44,14 @@ catalog:
   "@hyperledger/anoncreds-shared": 0.3.1
   "@openwallet-foundation/askar-react-native": 0.3.2
   "@openwallet-foundation/askar-shared": 0.3.2
+  "@sd-jwt/decode": ^0.14.0
+  "@sd-jwt/present": ^0.14.0
   buffer: 6.0.3
-  "@astronautlabs/jsonpath": ^1.1.2
+  credo-ts-didweb-anoncreds: 0.0.1-alpha.14
+  fast-text-encoding: ^1.0.6
+  isomorphic-webcrypto: npm:@sphereon/isomorphic-webcrypto
+  rxjs: ^7.8.1
+  query-string: ^8.1.0
   
   # React Native
   "@react-native-community/netinfo": ~11.4.1
@@ -89,6 +91,7 @@ catalog:
   expo-image-manipulator: ~13.1.7
   expo-linear-gradient: ~14.1.5
   expo-linking: ~7.1.6
+  expo-modules-autolinking: ~2.1.13
   expo-router: ~5.1.2
   expo-secure-store: ~14.2.3
   expo-splash-screen: ~0.30.9
@@ -102,6 +105,7 @@ catalog:
   # Tamagui
   "@tamagui/animations-react-native": 1.129.12
   "@tamagui/font-inter": 1.129.12
+  "@tamagui/portal": 1.129.12
   "@tamagui/shorthands": 1.129.12
   "@tamagui/themes": 1.129.12
   "@tamagui/toast": 1.129.12
